requirements_versions.txt
asgiref==3.8.1
attrs==25.3.0
audioread==3.0.1
babel==2.17.0
biopython==1.85
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.4
csvw==3.5.1
decorator==5.2.1
defusedxml==0.7.1
Django==5.2.3
django-browser-reload==1.18.0
django-cleanup==9.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
dlinfo==2.0.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.6.1
filelock==3.18.0
fsspec==2025.5.1
idna==3.10
inflection==0.5.1
isodate==0.7.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
language-tags==1.2.0
lazy_loader==0.4
Levenshtein==0.27.1
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
networkx==3.5
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.3.1
openai-whisper==20240930
packaging==25.0
phonemizer==3.3.0
platformdirs==4.3.8
pooch==1.8.2
PyAudio==0.2.14
pycparser==2.22
pydub-ng==0.2.0
PyJWT==2.9.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-Levenshtein==0.27.1
python-magic==0.4.27
python3-openid==3.2.0
pywatchman==3.0.0
PyYAML==6.0.2
RapidFuzz==3.13.0
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rfc3986==1.5.0
rpds-py==0.25.1
scikit-learn==1.7.0
scipy==1.15.3
segments==2.3.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
soundfile==0.13.1
soxr==0.5.0.post1
sqlparse==0.5.3
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.9.0
torch==2.7.1
tqdm==4.67.1
triton==3.3.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.4.0
